test: force atlas2 test
